Move `relaxed_atomic` into a separate file
